requirements.txt
py-cord==2.6.1
python-dotenv==1.0.1
aiohttp==3.11.11
openai==1.59.3
aiosqlite==0.20.0
httpx==0.28.1
Pillow==11.0.0
aiofiles==24.1.0
pytest==8.3.4
pytest-asyncio==0.25.1
pytest-mock==3.14.0
